requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
bcrypt

music21>=8.0.0
geoip2
user-agents
